Move the `database` directory at root to `src`.
