+++ src/database/models/Result.ts
