+++ src/index.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import helmet from 'helmet';
-import * as database from '../database';
+import * as database from './database';
 import { SERVER_PORT, UPLOAD_LIMIT } from '../markr.config';
 import * as routes from './routes';
